Fully commit to HCL and Subiquity - Increase Version to 3
Add http/user-data as the static configuration file to drive
the Subiquity-based initialization process.

Tweaked the scripting:
cleanup, minimize - added echos for long-running processes
virtualbox - removed code for obsolete 4.3.10 version
disable-upgrade-release - disables major release upgrades
always-update-existing - post-initialization, bring packages current
update - separated out into the above two scripts, but kept
         the major release upgrade unavailable for now.

Make it clear that desktop is an upgrade done post-initialization
and separate out the variables related to it from the base install.
The behavior of `packer build -var-file` will ensure the correct overrides
are applied.

The main build process is now done via `packer build`. There is no
default release, the user must specify one of the root HCL files on
the command line, and may specify to install the desktop role
as well.  An artifact version is also now an input to the build
process.  While the README suggest a timestamp+counter format
the value is simply used as-provided.

The specific behavior of packer here is to scan the repo for all
pkr.hcl files in the build-ubuntu subdirectory (specified as the
first argument on the command line) and load their contents into
memory.  It finds the build block (present in build.pkr.hcl) and
the child sources array to determine what to build - filtering
on the value of the -only option if one is supplied.  The names
of the sources resolve to the source blocks in the source-*pkr.hcl
files. While those blocks have a large amount of similar and even
identical content for now they only share variable values, their
key structure, which is highly Ubuntu-specific, is copy-and-paste.
In particular, the "boot_command" and related keys that specify
how packer is supposed to interact with the ISO.  Defaults for these
are specified in the vars-*.pkr.hcl files but can be overridden
by entries in the individual source-* file or the command-line
-var-file entry or entries for the release version and desktop mode,
or on the `packer build` command line.

Added a development-environments directory for developers to
record which version of software packages were used to perform
various builds.

In addition to the changes and additions described above
a number of items have been removed.

Removal of confirmed build environments for vmware and parallels
due to non-usage by me.  I hope others step in and fill in the gaps.
I left hcl files in place but did not test them.

Removal of versions < 16.04 as they are not Subiquity-based
nor are they supported by Canonical any longer.

Removal of floppy/* and http/preseed* as they are not part
of the Subiquity initialization process.

Removal of bin/* & Makefile - using packer directly for now

Removal of CHANGELOG.md due to lack of maintenance (and git log)

The above serves my present needs.  Future considered work
includes having at least http/user-data be dynamically constructed,
which probably will require reconstituting at least some pieces of the
/bin directory to facilitate and encapsulate.  A similar dynamic
of the source*.pkr-hcl files, deriving them from a shared template,
also is being considered.  There are no plans for me to using a
provider other than VirtualBox.  Due to all of this I put this out
as a fork and do not intend to submit a pull request upstream.  If
they, or others, wish to leverage the HCL work and incorporate that
back upstream I would be thrilled.  I don't have a desire to maintain
a full fork but felt these improvements were worthwhile to do.