Improve artifact naming UI
Set default value of "version" to null to force callers to choose
their desired version suffix on the artifact name.

Add a interactive_mode variable, set to server by default but toggled
to desktop in the add-desktop json, which populates the middle section
of the artifact name.  The vm_name from the main versioned ubuntu file
provide the name prefix.