Further removals and reworking of the user interface for the repo
In particular note that the user-data file is hard-coded and so
remove (or make local as appropriate) related variables.  In
particular SSH, though now I wonder whether/how the install user
and the vagrant.sh script should play together.  For now only
the default vagrant user is created (which bypasses the creation
block and just leaves general sudo permissions and ssh key stuff.